pyproject.toml
[project]
name = "ae-hw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openevolve",
    "python-dotenv>=1.1.0",
]

[tool.uv.sources]
openevolve = { git = "https://github.com/codelion/openevolve.git" }
